moved all methods related to portalConfigFinder to common
